lift ArrayOps into lms.collection
